LPD-62568 Add edit tags action to bulk actions dropdown
